+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes
 * [UI]: Fixed issue caused when trying to use the command line linker to export a large number of samples from a project.
 * [Developer]: Fixed issue caused when trying to create a new pipeline with no parameters.
 * [UI]: Moved "Members", "Groups", "Associated Projects", "Reference Files" to a new "Settings" tab in projects.
+* [UI]: Fixed bug where collaborators on a project had access to the sample tools. These buttons are now hidden from collaborators.
 * [Developer]: Removed `thymeleaf-extras-conditionalcomments`. Not used anymore.
 
 0.14.0 to 0.15.0